% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrle-package.R
\docType{package}
\name{myrle-package}
\alias{myrle}
\alias{myrle-package}
\title{myrle: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Hugo Gruson \email{hugo.gruson+R@normalesup.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID}) [copyright holder]

}
\keyword{internal}
